Steps toward DPLL algorithm
Added dpll_formula, prune_stack, and partial_assignment.
dpll_formula holds a copy of problem's graph it can modify, and it reversably prunes vertices and relaxes the graph.
prune_stack stores prune actions to a stack that can be used to reverse prune and relaxation actions.
partual_assignment is the same as assignment but with a tribool instead of bool.
Moved assignment into sat/assignment.